[142] [bug解决] 修改订单Id为Long导致的一些错误 && 一些页面的优化

+++ server/applet-server/src/main/java/com/imengli/appletServer/controller/OrderController.java
@@ -64,7 +64,7 @@ public class OrderController {
 
     @RequestMapping("/confirmCollection")
     public ResultDTO confirmCollection(@RequestParam String token,
-                                       @RequestParam Integer orderId) {
+                                       @RequestParam Long orderId) {
         return orderService.confirmCollection(token, orderId);
     }
